Converted floats to ints in format input lists for output {:d}

+++ WindowElements.py
@@ -847,12 +847,12 @@ class RightsWindow(TableWindow):
             
             rightName = Label()
             rightName['text'] = right[0]
-            rightName['sizen'] = '{:d} 2'.format((size[0] - 8) * 0.2)
+            rightName['sizen'] = '{:d} 2'.format(int((size[0] - 8) * 0.2))
             row.append(rightName)
             
             rightDescription = Label()
             rightDescription['text'] = right[1]
-            rightDescription['sizen'] = '{:d} 2'.format((size[0] - 8) * 0.8)
+            rightDescription['sizen'] = '{:d} 2'.format(int((size[0] - 8) * 0.8)
             row.append(rightDescription)
             
         self.setTable(rows, size[0] // 3, (7, int(size[0] * 0.2), int(size[0] * 0.8)),
